Add support for registration of discovered SPI PersistenceProvider
Currently the OSGi TCK requires that PersistenceProviders are registered
as a service (see https://github.com/osgi/osgi/issues/734 ).

Even though PaxJPA does not require this as it includes an own extender,
we can offer an option to register discovered providers as services to
make the TCK pass in this case.